+++ index.js
@@ -3,10 +3,11 @@ const castoeStreamer = require('./src/index.js');
 const castoeConsole = new castoeStreamer.Transports.Console({
 	name: 'Castoe Console Testing',
 	traceFile: true,
-	date: 'LTS',
-	showType: true
+	showType: false
 });
 
-castoeConsole.send('Testing');
+// castoeConsole.send('Testing');
 
-castoeConsole.send(true);
+// castoeConsole.send(true);
+
+castoeStreamer.Functions.rename('./src/TestFile.md', './Testing.md');